requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
datetime==4.4
pandas==1.3.5
requests==2.27.1
#
#json
#re